requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
dj-database-url==2.1.0
Django==5.0.4
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
environ==1.0
idna==3.7
PyJWT==2.8.0
requests==2.31.0
resend==0.8.0
sqlparse==0.5.0
typing_extensions==4.11.0
urllib3==2.2.1
